+++ client/package.json
@@ -10,6 +10,7 @@
     "preview": "vite preview"
   },
   "dependencies": {
+    "@reduxjs/toolkit": "^1.9.5",
     "@types/node": "^20.3.3",
     "axios": "^1.4.0",
     "react": "^18.2.0",
