+++ client/src/common/api/api.ts
@@ -1,12 +1,12 @@
 import { loginData, signUpData } from '../types/authTypes';
-import { authInstance, userInstance } from './instance';
+import { authInstance, instance, loginInstance } from './instance';
 
 export const authApi = {
-  login: (data: loginData) => authInstance.post('/users/login', data),
-  logout: () => authInstance.post('/users/logout'),
-  signup: (data: signUpData) => authInstance.post('/users/signup', data),
+  login: (data: loginData) => loginInstance.post('/users/login', data),
+  logout: () => instance.post('/users/logout'),
+  signup: (data: signUpData) => instance.post('/users/signup', data),
 };
 
 export const userApi = {
-  getUser: (userId: number) => userInstance.get(`/user/${userId}`),
+  getUser: (userId: string) => authInstance.get(`/user/${userId}`),
 };
